+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - When expired, they'll report any skipped messages that haven't been reported before.
   - This should help with long-running processes that could use increasing amounts of memory in the rare case that they continually generate both large and unique log messages (actually unique `stream_ids`).
   - The new configurable parameters introduced for this feature are: `expire_check_sec`, `expire_offset_sec` and `expire_msg`.
+- `stream_id_max_len` option can be used to limit `stream_id` strings to a maximum length.
 
 ## [1.2.2] - 2022-12-22